extracting append_dir(const char *, const char *)
The purpose of append_dir is to complete a path adding to it a new
directory.
Also we can use it to ensure that a path has a traling slash, i.e :

  appending "b" to "a", generate "a/b/"

The function ensures that the return path has only *one* trailing
slash, i.e:

  appending "b///" to "a/" generate "a/b/"

but it doesn't ensure that the return is a valid path, i.e:

  appending "c//" to "a///b" generate "a///b/c/"

This should be a rare case since we work with real paths.